tests(thesis): split unit tests
